% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.plot.R
\name{weight.plot}
\alias{weight.plot}
\title{Generates Plots of weights by group}
\usage{
weight.plot(fit.object, group.names, group.list, x.s)
}
\arguments{
\item{fit.object}{The object that is returned by the bgwqs.fit function}

\item{group.names}{A string vector containing the name of each group included in the BGWQS regression. Will be used for plot titles.}

\item{group.list}{A list, each item in the list being a string vector of variable names for one component group.}

\item{x.s}{A vector of the number of components in each index.}
}
\value{
A plot for each group of the BGWQS regression
}
\description{
This function takes the object created by the bgwqs.fit function and a vector of group names and generates a
random forest variable importance plot for each group. The weights in each group are listed in descending order.
}
